// ui/src/navigation/enums.ts
export const enum RootRoutes {
  Barber = 'Barber',
  Chat = 'Chat',
  Login = 'Login',
  Signup = 'Signup',
  Welcome = 'Welcome',
  Main = 'Main',
  NotFound = 'NotFound',
}

export const enum MainRoutes {
  Home = 'Home',
  Search = 'Search',
  Messages = 'Messages',
  SettingsStack = 'SettingsStack',
}

export const enum SettingsRoutes {
  Settings = 'Settings',
  Contact = 'Contact Us',
  Account = 'Account',
  Notifications = 'Notifications',
  Payment = 'Payment Details',
  Preferences = 'Preferences',
}